Mark IT as IT so that maven can run them independently from unit tests
