+++ pom.xml
@@ -20,13 +20,6 @@
         <java.version>14</java.version>
     </properties>
 
-    <!--    <parent>-->
-    <!--        <groupId>org.springframework.boot</groupId>-->
-    <!--        <artifactId>spring-boot-starter-parent</artifactId>-->
-    <!--        <version>2.3.0.RELEASE</version>-->
-    <!--        <relativePath/>-->
-    <!--    </parent>-->
-
     <repositories>
         <repository>
             <id>oss.sonatype.org</id>
@@ -52,22 +45,6 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-validation</artifactId>
         </dependency>
-        <!--        <dependency>-->
-        <!--            <groupId>javax.ejb</groupId>-->
-        <!--            <artifactId>javax.ejb-api</artifactId>-->
-        <!--            <version>3.2.2</version>-->
-        <!--        </dependency>-->
-        <!--        <dependency>-->
-        <!--            <groupId>javax.persistence</groupId>-->
-        <!--            <artifactId>javax.persistence-api</artifactId>-->
-        <!--            <version>2.2</version>-->
-        <!--        </dependency>-->
-        <!--        <dependency>-->
-        <!--            <groupId>javax.validation</groupId>-->
-        <!--            <artifactId>validation-api</artifactId>-->
-        <!--            <version>2.0.1.Final</version>-->
-        <!--        </dependency>-->
-
         <!-- LOMBOK -->
         <dependency>
             <groupId>org.projectlombok</groupId>
@@ -109,6 +86,7 @@
                 </exclusion>
             </exclusions>
         </dependency>
+
 
     </dependencies>
 
@@ -171,10 +149,6 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <reuseForks>false</reuseForks>
-                    <forkCount>1</forkCount>
-                </configuration>
                 <version>3.0.0-M5</version>
             </plugin>
